+++ index.js
@@ -1,9 +1,11 @@
-var defaultTags = require('./app/tag_parser').defaultTags;
-var xejs = require('./app/renderer');
+"use strict";
+
+const defaultTags = require('./app/tag_parser').defaultTags;
+const xejs = require('./app/renderer');
 
 
 function setOptions(renderingOptions, args) {
-    var tokens = defaultTags;
+    let tokens = defaultTags;
     if (renderingOptions.tokens) tokens = tokens.concat(renderingOptions.tokens);
     return {
         openTagEJS: "<%- ",
@@ -29,10 +31,10 @@ function renderFile(file, renderingOptions, args, done) {
         renderingOptions = {};
     }
 
-    var options = setOptions(renderingOptions, args);
+    const options = setOptions(renderingOptions, args);
 
-    var res = null;
-    var err = null;
+    let res = null;
+    let err = null;
     try {
         res = xejs.renderFile(file, options);
     } catch (e) {
@@ -58,11 +60,11 @@ function renderString(content, renderingOptions, args, done) {
         renderingOptions = {};
     }
 
-    var options = setOptions(renderingOptions, args);
-    var includePath=renderingOptions.includePath;
+    const options = setOptions(renderingOptions, args);
+    const includePath=renderingOptions.includePath;
 
-    var res = null;
-    var err = null;
+    let res = null;
+    let err = null;
     try {
         res = xejs.renderString(content, options, includePath);
     } catch (e) {